feat: aula01 projeto01 part 01
the first feature of this class, in it, we explore
the simple structure of a C# main method.

disclaimer: this is being done for learning
purposes, with .NET 6 and beyond, Microsoft has
adopted the concept of 'low code', meaning that
the namespace, class and main parts of Program.cs
don't need to be written out for the project to
compile.